Re-organized Stan's example into one folder. Simplified antools3.py into readspec.py, which is now just a code to read spectra from different kinds of fits files. Edited Stan's code to account for the changes from antools3 to readspec.
