+++ Example/TEST_PHEW_LITE/spec-57415-KP084317N212806V01_sp01-217.fits
